Removed mmap(2) support.  According to the mmap(2) man page:
MAP_PRIVATE

Create a private copy-on-write mapping.  Stores to the region do not
affect the original file.  It is unspecified whether changes made to
the file after the mmap call are visible in the mapped region.

NSD depends on the database not changing after opening, so mmap(2) is
not guaranteed to work.  Also, with plugins, the data is changed by
NSD.  Although MAP_PRIVATE prevents these changes from being written
to the database, this is still kind of "hacky".

So mmap(2) support has been removed.


git-svn-id: svn+ssh://open.nlnetlabs.nl/svn/nsd/trunk@696 a26ef69c-88ff-0310-839f-98b793d9c207